Compatibility with modern Expanded Storage
Changed some lines of code so that the mod is compatible with newer versions of Expanded Storage. If compiling doesn't work, try running this command in cmd:
```
gradlew -Dorg.gradle.java.home="C:\Program Files\Java\jdk-17
```
The mod won't crash when using newer versions of Expanded Storage (which apparently got deleted from curseforge along with the older one), but I don't think it places horizontal double chests correctly when a contraption is stopped. The fix is a few months old, so I don't remember a lot other than that it works.